[#56] chore: 파일들을 알맞게 grouping하고 파일을 분리함
