+++ iOS/Todo/Todo/ViewControllers/CardViewController.swift
@@ -85,27 +85,6 @@ final class CardViewController: UIViewController {
     }
 }
 
-final class CancelButton: UIButton {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configure()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        configure()
-    }
-    
-    private func configure() {
-        let cancelImage = UIImage(systemName: "x.circle.fill")
-        setImage(cancelImage, for: .normal)
-        imageView?.contentMode = .scaleAspectFill
-        imageEdgeInsets = UIEdgeInsets(top: 30, left: 30, bottom: 30, right: 30)
-        tintColor = .systemRed
-        translatesAutoresizingMaskIntoConstraints = false
-    }
-}
-
 final class CreateButton: UIButton {
     override init(frame: CGRect) {
         super.init(frame: frame)
@@ -124,74 +103,5 @@ final class CreateButton: UIButton {
         imageEdgeInsets = UIEdgeInsets(top: 30, left: 30, bottom: 30, right: 30)
         tintColor = .systemBlue
         translatesAutoresizingMaskIntoConstraints = false
-    }
-}
-
-final class TitleField: UITextField {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configure()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        configure()
-    }
-    
-    private func configure() {
-        translatesAutoresizingMaskIntoConstraints = false
-        font = UIFont.boldSystemFont(ofSize: 30)
-        placeholder = "Title"
-    }
-    
-    private static let padding = UIEdgeInsets(top: 0, left: 7, bottom: 0, right: 0)
-    override func textRect(forBounds bounds: CGRect) -> CGRect {
-        return bounds.inset(by: TitleField.padding)
-    }
-    
-    override func placeholderRect(forBounds bounds: CGRect) -> CGRect {
-        return bounds.inset(by: TitleField.padding)
-    }
-    
-    override func editingRect(forBounds bounds: CGRect) -> CGRect {
-        return bounds.inset(by: TitleField.padding)
-    }
-}
-
-final class ContentView: UITextView {
-    static let placeHolderString = "Add a message what to do"
-    
-    override init(frame: CGRect, textContainer: NSTextContainer?) {
-        super.init(frame: frame, textContainer: textContainer)
-        configure()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        configure()
-    }
-    
-    private func configure() {
-        translatesAutoresizingMaskIntoConstraints = false
-        configureText()
-        configureBorder()
-        configureInsets()
-    }
-    
-    private func configureText() {
-        font = UIFont.boldSystemFont(ofSize: 20)
-        text = ContentView.placeHolderString
-        textColor = .placeholderText
-    }
-    
-    private func configureBorder() {
-        layer.borderWidth = 0.9
-        layer.borderColor = UIColor.placeholderText.cgColor
-        layer.cornerRadius = 13
-    }
-    
-    private static let padding = UIEdgeInsets(top: 20, left: 10, bottom: 0, right: 0)
-    private func configureInsets() {
-        textContainerInset = ContentView.padding
     }
 }
