+++ iOS/Todo/Todo/Views/CardVIewController's/ContentView.swift
@@ -1,0 +1,46 @@
+//
+//  ContentView.swift
+//  Todo
+//
+//  Created by kimdo2297 on 2020/04/11.
+//  Copyright © 2020 Jason. All rights reserved.
+//
+import UIKit
+
+final class ContentView: UITextView {
+    static let placeHolderString = "Add a message what to do"
+    
+    override init(frame: CGRect, textContainer: NSTextContainer?) {
+        super.init(frame: frame, textContainer: textContainer)
+        configure()
+    }
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        configure()
+    }
+    
+    private func configure() {
+        translatesAutoresizingMaskIntoConstraints = false
+        configureText()
+        configureBorder()
+        configureInsets()
+    }
+    
+    private func configureText() {
+        font = UIFont.boldSystemFont(ofSize: 20)
+        text = ContentView.placeHolderString
+        textColor = .placeholderText
+    }
+    
+    private func configureBorder() {
+        layer.borderWidth = 0.9
+        layer.borderColor = UIColor.placeholderText.cgColor
+        layer.cornerRadius = 13
+    }
+    
+    private static let padding = UIEdgeInsets(top: 20, left: 10, bottom: 0, right: 0)
+    private func configureInsets() {
+        textContainerInset = ContentView.padding
+    }
+}
