+++ iOS/Todo/Todo/Views/CardVIewController's/FormatImageView.swift
@@ -1,0 +1,27 @@
+//
+//  FormatImageView.swift
+//  Todo
+//
+//  Created by kimdo2297 on 2020/04/11.
+//  Copyright © 2020 Jason. All rights reserved.
+//
+import UIKit
+
+final class FormatImageView: UIImageView {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configure()
+    }
+    
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        configure()
+    }
+    
+    private func configure() {
+        translatesAutoresizingMaskIntoConstraints = false
+        heightAnchor.constraint(equalToConstant: 25).isActive = true
+        widthAnchor.constraint(equalTo: heightAnchor, multiplier: 1).isActive = true
+        image = UIImage(systemName: "text.justify")!
+    }
+}
