+++ iOS/Todo/Todo/Views/ColumnViewController's/CardCell.swift
@@ -76,25 +76,6 @@ final class CardCell: UITableViewCell, ReusableView {
     }
 }
 
-final class FormatImageView: UIImageView {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configure()
-    }
-    
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-        configure()
-    }
-    
-    private func configure() {
-        translatesAutoresizingMaskIntoConstraints = false
-        heightAnchor.constraint(equalToConstant: 25).isActive = true
-        widthAnchor.constraint(equalTo: heightAnchor, multiplier: 1).isActive = true
-        image = UIImage(systemName: "text.justify")!
-    }
-}
-
 final class CardTitleLabel: UILabel {
     override init(frame: CGRect) {
         super.init(frame: frame)
